ISSUE #10
- moved strategies to separate file `strategies.py`
- moved help functions to `utils.py`
- generate model functions shifted to `services` folder & subfiles
- read constants from `config.py`

Closes #10
